% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{CPdata}
\alias{CPdata}
\title{CPdata: Example dataset for aLBI package}
\format{
A data frame with [number of rows] rows and 11 columns:
\describe{
  \item{\code{A}}{[description of column A]}
  \item{\code{B}}{[description of column B]}
  \item{\code{C}}{[description of column C]}
  \item{\code{D}}{[description of column D]}
  \item{\code{E}}{[description of column E]}
  \item{\code{F}}{[description of column F]}
  \item{\code{G}}{[description of column G]}
  \item{\code{H}}{[description of column H]}
  \item{\code{I}}{[description of column I]}
  \item{\code{J}}{[description of column J]}
  \item{\code{Tx}}{[description of column Tx]}
}
}
\source{
[Include the source of the data if available]
}
\usage{
data(CPdata)
}
\description{
This dataset contains description of CPdata.
}
\examples{
data(CPdata)
head(CPdata)
}
