% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{ExData}
\alias{ExData}
\title{ExData: Example raw length dataset for aLBI package}
\format{
A data frame with 1 column:
\describe{
  \item{\code{Length}}{Sampled length data (cm)}
}
}
\source{
Data collected for fish stock assessment studies
}
\usage{
data(ExData)
}
\description{
This dataset contains description of ExData.
}
\examples{
data(ExData, package = "aLBI")
head(ExData)
}
