% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{lenfreq01}
\alias{lenfreq01}
\title{lenfreq01: Example dataset for aLBI package}
\format{
A data frame with [number of rows] rows and 2 columns:
\describe{
  \item{\code{Frequency}}{[description of column Frequency]}
  \item{\code{Length}}{[description of column Length]}
}
}
\source{
[Include the source of the data if available]
}
\usage{
data(lenfreq01)
}
\description{
This dataset contains description of lenfreq01.
}
\examples{
data(lenfreq01)
head(lenfreq01)
}
