% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{lenfreq02}
\alias{lenfreq02}
\title{lenfreq02}
\format{
A data frame.
}
\source{
Self-generated data.
}
\usage{
lenfreq02
}
\description{
This is another dummy length frequency data set used for testing purposes.
}
\keyword{datasets}
