% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{lenfreqM}
\alias{lenfreqM}
\title{lenfreqM: Example length-weight dataset for aLBI package}
\format{
A data frame with 2 columns: First one is month and second one is length
\describe{
  \item{\code{Months}}{Name of the Sampling Months}
  \item{\code{Length}}{Measured lengths of the sampled fish}
}
}
\source{
Data collected for fish stock assessment studies
}
\usage{
data(lenfreqM)
}
\description{
This dataset contains length data of multiple months
}
\examples{
data(lenfreqM, package = "aLBI")
head(lenfreqM)
}
