% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{LWdata}
\alias{LWdata}
\title{LWdata: Example length-weight dataset for aLBI package}
\format{
A data frame with 2 columns:
\describe{
  \item{\code{Length}}{Length of sampled fish (cm)}
  \item{\code{Weight}}{Weight of sampled fish (g)}
}
}
\source{
Data collected for fish stock assessment studies
}
\usage{
data(LWdata)
}
\description{
This dataset contains length and weight measurements for fish.
}
\examples{
data(LWdata, package = "aLBI")
head(LWdata)
}
