% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{CPdata}
\alias{CPdata}
\title{CPdata}
\format{
A data frame with 21 rows and 11 columns.
}
\source{
References: Cope, J. M., & Punt, A. E. (2009). Length‐Based Reference Points for Data‐Limited Situations: Applications and Restrictions. Marine and Coastal Fisheries, 1(1), 169–186. https://doi.org/10.1577/C08-025.1
}
\usage{
CPdata
}
\description{
This dataset contains information about the stock status Probability according to Cope & Punt, 2009.
}
\keyword{datasets}
