% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{CPdata}
\alias{CPdata}
\title{CPdata: Example dataset for aLBI package}
\format{
A data frame with 11 columns:
\describe{
  \item{\code{A}}{Probability values}
  \item{\code{B}}{Probability values}
  \item{\code{C}}{Probability values}
  \item{\code{D}}{Probability values}
  \item{\code{E}}{Probability values}
  \item{\code{F}}{Probability values}
  \item{\code{G}}{Probability values}
  \item{\code{H}}{Probability values}
  \item{\code{I}}{Probability values}
  \item{\code{J}}{Probability values}
  \item{\code{Tx}}{Target column compared with LM_ratio to pick probability values}
}
}
\source{
A decision table described by Cope and Punt (2009)
}
\usage{
data(CPdata)
}
\description{
This dataset contains description of CPdata.
}
\examples{
data(CPdata, package = "aLBI")
head(CPdata)
}
