% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{ExData}
\alias{ExData}
\title{ExData: Example raw lenght dataset for aLBI package}
\format{
A data frame with [number of rows] rows and 1 columns:
\describe{
  \item{\code{Length}}{[description of column Length measured]}
}
}
\source{
[Include the source of the data if available]
}
\usage{
data(ExData)
}
\description{
This dataset contains description of ExData.
}
\examples{
data(ExData)
head(ExData)
}
