% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{lenfreq01}
\alias{lenfreq01}
\title{lenfreq01}
\format{
A data frame.
}
\source{
Self-generated data.
}
\usage{
lenfreq01
}
\description{
This is a dummy length frequency data set used for testing purposes.
}
\keyword{datasets}
