% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{lenfreq01}
\alias{lenfreq01}
\title{lenfreq01: Example dataset for aLBI package}
\format{
A data frame with 2 columns:
\describe{
  \item{\code{Frequency}}{Observed individuals in each length class}
  \item{\code{Length}}{Upper value of each length class (cm)}
}
}
\source{
Data collected for fish stock assessment studies
}
\usage{
data(lenfreq01)
}
\description{
This dataset contains description of lenfreq01.
}
\examples{
data(lenfreq01, package = "aLBI")
head(lenfreq01)
}
