% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{lenfreq02}
\alias{lenfreq02}
\title{lenfreq02: Example dataset for aLBI package}
\format{
A data frame with [number of rows] rows and 2 columns:
\describe{
  \item{\code{Frequency}}{[description of column Frequency]}
  \item{\code{LengthClass}}{[description of column LengthClass]}
}
}
\source{
[Include the source of the data if available]
}
\usage{
data(lenfreq02)
}
\description{
This dataset contains description of lenfreq02.
}
\examples{
data(lenfreq02)
head(lenfreq02)
}
