% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{lenfreq02}
\alias{lenfreq02}
\title{lenfreq02: Example dataset for aLBI package}
\format{
A data frame with 2 columns:
\describe{
  \item{\code{Frequency}}{Observed individuals in each length class}
  \item{\code{LengthClass}}{Upper value of each length class (cm)}
}
}
\source{
Data collected for fish stock assessment studies
}
\usage{
data(lenfreq02)
}
\description{
This dataset contains description of lenfreq02.
}
\examples{
data(lenfreq02, package = "aLBI")
head(lenfreq02)
}
